testmysql/requirements.txt
mysql_connector_python==8.0.26
